WIP: Salvamento temporário antes de pull
